pyproject.toml
[project]
name = "candidate-matcher"
version = "1.0.0"
description = "An agentic candidate-job matcher system"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.12.0",
    "pydantic-ai>=1.47.0",
    "python-dotenv>=1.2.1",
    
    # LLM Providers
    "openai>=2.15.0",
        
    # UI
    "streamlit>=1.53.0",
    
    # Document Processing
    "pymupdf>=1.26.0",
    "beautifulsoup4>=4.14.0",
    "requests>=2.32.0",
    
    # PII Detection
    "transformers>=4.57.0",
    "torch==2.10.0; platform_system == 'Windows'",

    
    # Embeddings for Semantic Matching
    "sentence-transformers>=5.2.0",
    "scikit-learn>=1.4.0",
    
    # Formatting
    "pyyaml>=6.0.0"
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

